拆分FtpUtils类，解除common模块依赖Collect模块
