+++ Common/src/main/java/com/hzgc/common/ftp/FtpUtils.java
@@ -45,24 +45,6 @@ public final class FtpUtils implements Serializable {
     }
 
     /**
-     * @param pictureName determine the picture type based on the file name
-     * @return equals 0, it is a picture
-     * lager than 0, it is a face picture
-     */
-    public static int pickPicture(String pictureName) {
-        int picType = 0;
-        if (null != pictureName) {
-            String tmpStr = pictureName.substring(pictureName.lastIndexOf("_") + 1, pictureName.lastIndexOf("."));
-            try {
-                picType = Integer.parseInt(tmpStr);
-            } catch (Exception e) {
-                e.printStackTrace();
-            }
-        }
-        return picType;
-    }
-
-    /**
      * 根据文件上传至ftp的绝对路径获取到ipcID、TimeStamp、Date、TimeSlot
      *
      * @param fileName 文件上传至ftp的绝对路径，例如：/3B0383FPAG51511/2017/05/23/16/00/2017_05_23_16_00_15_5704_0.jpg
@@ -129,29 +111,6 @@ public final class FtpUtils implements Serializable {
         }
         return message;
     }
-
-    /**
-     * 通过上传文件路径解析到文件的ftp地址（ftp发送至kafka的key）
-     *
-     * @param filePath ftp接收数据路径
-     * @return 文件的ftp地址
-     */
-    public static String filePath2FtpUrl(String filePath) {
-        StringBuilder url = new StringBuilder();
-        String hostName = IPAddressUtils.getHostName();
-        //TODO 为编译通过，临时注释
-        /*Map<Integer, Integer> ftpPIDMap = FTP.getPidMap();
-        if (!ftpPIDMap.isEmpty()) {
-            Integer ftpPID = Integer.valueOf(ManagementFactory.getRuntimeMXBean().getName().split("@")[0]);
-            //LOG.info("ftp PID = " + ftpPID);
-            int ftpPort = ftpPIDMap.get(ftpPID);
-            url = url.append("ftp://").append(hostName).append(":").append(ftpPort).append(filePath);
-            return url.toString();
-        }
-        url = url.append("ftp://").append(hostName).append(":").append("none").append(filePath);*/
-        return url.toString();
-    }
-
 
     /**
      * 小图ftpUrl转大图ftpUrl
